Major reconciliation Synrad/Molflow
- Correction for density near absorbing/desorbing facets
- Added pressure, denisty, imp.rate, etc. to FormulaSettings
- Insertgeometry: delayed formula parsing
- Outgassing Map: 1E-9 -> 0.999999999
- Profile plotter SetWorker() routine
- Additional Insertgeometry types
- Cycle selections
- FrameMove can be disabled
- More detailed Wait() routine (displays status of workers)

+++ FacetCoordinates.cpp
@@ -257,7 +257,7 @@ void FacetCoordinates::RebuildList() {
 	
 	char tmp[128];
 
-	for(size_t i=0;i<lines.size();i++) {
+	for(int i=0;i<(int)lines.size();i++) {
 		
 		sprintf(tmp,"%d",i+1);
 		facetListC->SetValueAt(0,i,tmp);
