+++ SMP/ShMemory.cpp
@@ -28,7 +28,7 @@ void PrintLastErrorText( LPTSTR suff );
 // CreateDataport: Create a block of shared memory
 // --------------------------------------------------------------------------
 
-Dataport *CreateDataport(char *name,long size)
+Dataport *CreateDataport(char *name,size_t size)
 {
    Dataport  *dp;
    
@@ -55,7 +55,7 @@ Dataport *CreateDataport(char *name,long size)
 			      	NULL,                   // no security 
 				      PAGE_READWRITE,         // to allow read & write access
 				      0,     
-				      size,                   // file size     
+				      (DWORD)size,                   // file size     
 				      name);                  // object name      
 
    if( GetLastError()==ERROR_ALREADY_EXISTS ) {
@@ -107,7 +107,7 @@ Dataport *CreateDataport(char *name,long size)
 // OpenDataport: Connect to an existing block
 // --------------------------------------------------------------------------
 
-Dataport *OpenDataport(char *name,long size)
+Dataport *OpenDataport(char *name,size_t size)
 {
    Dataport *dp;
 
@@ -133,7 +133,7 @@ Dataport *OpenDataport(char *name,long size)
 		       		NULL,                   // no security 
 			      	PAGE_READWRITE,         // to allow read & write access
 		      		0,     
-		      		size,                   // file size     
+		      		(DWORD)size,                   // file size     
 			      	name);                  // object name      
    
    if( GetLastError()!=ERROR_ALREADY_EXISTS ) {
